requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
mysql-connector==2.2.9
mysql-connector-python==9.1.0
